feat: added application architecture for project
