+++ src/components/footer/footer.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Footer = () => {
+    return (
+        <footer>
+            FOOTER
+        </footer>
+    );
+};
+
+export default Footer;
