+++ src/components/header/header.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Header = () => {
+    return (
+        <header>
+            HEADER
+        </header>
+    );
+};
+
+export default Header;
